+++ src/app/__tests__/get-task-usecase.spec.ts
@@ -2,7 +2,7 @@ import { PrismaClient } from '@prisma/client';
 import { TaskQueryService } from 'src/infra/db/query-service/task-query-service';
 import { MockedObjectDeep } from 'ts-jest/dist/utils/testing';
 import { mocked } from 'ts-jest/utils';
-import { GetTaskUseCase } from '../get-task-usecae';
+import { GetTaskUseCase } from '../get-task-usecase';
 import * as faker from 'faker';
 import { TaskDTO } from '../query-service-interface/task-query-service';
